feat: enhance MainWebview navigation and update DefectForm and UserStoryForm styling
Introduces a new reference in MainWebview to track the origin of detail screens, allowing for improved back navigation to the Home section. Updates the handling of selected items to reset the detail origin appropriately. Additionally, refactors the DefectForm and UserStoryForm components to utilize a ResizableDescription component for better description handling and styling, improving overall user experience and visual consistency.
